+++ rvtools2azmigrate/config.py
@@ -1,0 +1,11 @@
+VINFO_SHEET_NAME = "vInfo"
+"""Name of the Excel sheet to look for VM data"""
+
+STORAGE_COLUMN_NAME = "Provisioned"
+"""Name of the column to look for VM storage capacity"""
+
+MIB_TO_MB_CONVERSION_FACTOR = 1.04858
+"""Conversion factor to convert MiB to MB"""
+
+DEFAULT_OS_NAME = "Windows Server 2019 Datacenter"
+"""Use this OS Name when no data is available to determine the OS"""