+++ HISTORY.md
@@ -1,5 +1,11 @@
 # History
 
+## 0.1.1 (2023-02-03)
+
+* Filter Powered-off VMs
+* Automatic MiB/MB detection
+* Intermediate `RvToolsVM` and `AzMigrateVM` objects
+
 ## 0.1.0 (2022-12-22)
 
 * First release on GitHub.